+++ hospital/src/hospital/EventoFinSimulacion.java
@@ -5,8 +5,9 @@ public class EventoFinSimulacion extends Evento {
     /**
      * Constructor de la clase EventoFinSimulacion
      * @param tiempo instante de tiempo en el que se encuentra la simulacion al momento de generar el evento
+     * @param tipo tipo de evento
      */
-    public EventoFinSimulacion(float tiempo,int tipo){
+    public EventoFinSimulacion(double tiempo,int tipo){
         super(2,tiempo,new Item(tiempo,tipo));
     }
